+++ test/case/proxy-cookies/config.js
@@ -4,8 +4,8 @@ module.exports = {
   },
   proxy: {
     localRemote: {
-      'http://yylive-ssr.yy.com/': 'http://127.0.0.1:5000/'
+      'http://yylive-ssr.testhost.com/': 'http://127.0.0.1:5000/'
     },
-    homePage: 'http://yylive-ssr.yy.com/'
+    homePage: 'http://yylive-ssr.testhost.com/'
   }
 }