+++ test/case/proxy/config.js
@@ -8,20 +8,20 @@ const config = {
   proxy: {
     port: 8887,
     localRemote: {
-      'http://web.yy.com/': 'http://127.0.0.1:5000/',
-      'http://www.yy.com/api/mock': 'http://127.0.0.1:5000/api/mock',
+      'http://www.testhost.com/': 'http://127.0.0.1:5000/',
+      'http://www.testhost.com/api/mock': 'http://127.0.0.1:5000/api/mock',
       'http://web.yystatic.com/': 'http://127.0.0.1:5000/'
     },
-    homePage: 'http://www.yy.com/web/1/'
+    homePage: 'http://www.testhost.com/web/1/'
   },
   commit: {
-    hostname: '//web.yystatic.com',
-    staticHost: '//web.yystatic.com',
-    mainHost: '//www.yy.com/web'
+    hostname: '//www.testhost.com',
+    staticHost: '//www.testhost.com',
+    mainHost: '//www.testhost.com/web'
   }
 }
 config.proxy.localRemote[
-  `http://www.yy.com/web/${PROJECT_NAME}`
+  `http://www.testhost.com/web/${PROJECT_NAME}`
 ] = `http://127.0.0.1:5000/project/${PROJECT_NAME}/${PLATFORM}/html`
 
 module.exports = config